pyproject.toml
[project]
name = "quiz-project"
version = "0.1.0"
description = ""
authors = [
    {name = "nevertheless0404",email = "nevertheless.hi@gmail.com"}
]
requires-python = ">=3.12,<4.0"
dependencies = [
    "fastapi[all] (>=0.115.11,<0.116.0)",
    "uvicorn (>=0.34.0,<0.35.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "sqlalchemy (>=2.0.39,<3.0.0)",
    "alembic (>=1.15.1,<2.0.0)",
    "asyncpg (>=0.30.0,<0.31.0)",
    "redis (>=5.2.1,<6.0.0)",
    "aioredis (>=2.0.1,<3.0.0)",
    "fastapi-limiter (>=0.1.6,<0.2.0)",
    "pydantic (>=2.10.6,<3.0.0)",
    "django (>=5.1.7,<6.0.0)",
    "djangorestframework (>=3.15.2,<4.0.0)",
    "python-dotenv (>=1.0.1,<2.0.0)",
    "greenlet (>=3.1.1,<4.0.0)",
    "djangorestframework-simplejwt (>=5.5.0,<6.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
